refactor #95: dto 컨벤션 적용 및 코드 정리
